Resolve pie chart todos
